Move crypto extensions into their own directory
