+++ packages/transmute-did/src/lib/cryptoSuites/ellipticExtensions/__tests__/ellipticExtensions.spec.js
