+++ packages/transmute-did/src/lib/cryptoSuites/ellipticExtensions/index.js
