+++ packages/transmute-did/src/lib/cryptoSuites/ethereumExtensions/__tests__/eth.spec.js
