+++ packages/transmute-did/src/lib/cryptoSuites/ethereumExtensions/index.js
