+++ packages/transmute-did/src/lib/cryptoSuites/openpgpExtensions/did/ocap/json-ld/__tests__/openpgp-ocap-ld.spec.js
@@ -6,7 +6,7 @@ const { getMessagePayload } = require('../../../../cryptoHelpers');
 
 const { openpgpSignJson, openpgpVerifyJson } = require('../index');
 
-const pack = require('../../../../../../../package.json');
+const pack = require('../../../../../../../../package.json');
 
 const openpgpVersion = `openpgp@${pack.dependencies.openpgp}`;
