+++ packages/transmute-did/src/lib/cryptoSuites/shamirExtensions/__tests__/shamir.spec.js
