+++ packages/transmute-did/src/lib/cryptoSuites/shamirExtensions/index.js
