+++ packages/transmute-did/src/lib/cryptoSuites/sodiumExtensions/__tests__/sodiumExtensions.spec.js
