+++ packages/transmute-did/src/lib/cryptoSuites/sodiumExtensions/index.js
